requirements.txt
bidict==0.21.2
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
flask-mongoengine==1.0.0
Flask-SocketIO==5.0.0
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.22.1
pymongo==3.11.2
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.3
Werkzeug==1.0.1
WTForms==2.3.3
